Rescale images during loading
The stretch, fit, and fill image modes often require that a large
image be scaled down to fit. In that case, it is often more faster
(and uses less peak memory) to avoid loading the full image, and
instead perform the scaling as the image is being loaded, with
`gdk_pixbuf_new_from_stream_at_scale`.