Remove redundant redraw request when scale changes
When an output's geometry changes, or a new output is introduced,
the only sensible thing for the compositor to do is send
xdg_output::scale, etc. before sending the zwlr_layer_shell_v1.configure
event which requests a surface resize. Therefore, we do not need
to redraw in response to the xdg_output::scale event.