Set default background color to #3f3f3fff
This ensures that the background is always opaque.

This is a significant change; before, if no background color was
specified, images were drawn onto an initially transparent buffer,
leaving some of the content behind swaybg visible if the image was
alpha transparent or if the 'fit' or 'center' scaling modes were used
and the image aspect ratio did not match, or the image was not large
enough. As the purpose of a wallpaper/background program is to draw
the pixels behind all other interface items, this transparency is
neither a required nor a very useful behavior.

The new color is close to the result of blending the old default
background color (#00000000 = transparent) with the color Sway uses
to clear the background ([0.25,0.25,0.25,1.0] = mid gray). Users of
other compositors may need to adjust their images or specify the
--color flag on the command line.