Rescale scalable images as they are loaded
This change uses GdkPixbufLoader to rescale scalable (SVG) images as
they are being loaded.

When an image is drawn on multiple outputs, it is loaded with the maximum
size over all the sizes at which the image is drawn on the different
outputs. If the outputs render the image at different scales, then on
some output the image will be scaled twice, which can reduce the
rendering quality. Such quality loss can be avoided by configuring swaybg
to choose image files that have already been scaled in advance.

The Cairo image loading code is mostly unaffected.